add default `callback query handler`
